requirements.txt
streamlitopenailangchainSQLAlchemypsycopg2-binarypython-dotenv